requirements.txt
asgiref==3.2.10
Django==3.1.2
djangorestframework==3.12.1
djangorestframework-simplejwt==4.4.0
jsonfield==3.1.0
pkg-resources==0.0.0
PyJWT==1.7.1
pytz==2020.1
sqlparse==0.4.1
